+++ NEWS.md
@@ -9,6 +9,8 @@
   copying `crayon` functions
   ([@seulki-choi](https://stackoverflow.com/users/7788015/seulki-choi),
   @gaborcsardi)
+* [#100](https://github.com/brodieG/diffobj/issues/100): R_useDynamicSymbols
+* [#105](https://github.com/brodieG/diffobj/issues/105): Copyright tweaks
 
 ## v0.1.6