+++ NEWS.md
@@ -12,6 +12,8 @@
 * [#100](https://github.com/brodieG/diffobj/issues/100): R_useDynamicSymbols
 * [#97](https://github.com/brodieG/diffobj/issues/97): 2D Guidelines fixes for
   data.table, tibble
+* [#96](https://github.com/brodieG/diffobj/issues/96): Warnings when comparing
+  large data tables.
 * [#105](https://github.com/brodieG/diffobj/issues/105): Copyright tweaks
 
 ## v0.1.6